cequiz-backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-stubs==1.9.0
